Move some files to new cli folder
