+++ ryebot/bot/cli/wiki_manager.py
